fix stream reader
There were 2 cases that the live query / stream reader was unable to
handle:

1. ':' keep alive messages
2. multiple data lines in a single SSE 'packet'

Both of these issues would cause issues when attempting to JSON decode
an SSE packet. To add insult to injury, there was a bug where we would
not close the stream on such an error. This would result in connections
stacking up.

This commit addresses these issues by:

1. Moving JSON parsing into readStream and ensuring
   that we close the stream on any kind of error.
2. Ignoring messages that are not 'data:' -- namely ':'